Added factor 10 in RCD2H
Added factor 10 to line 31 since the values in the original paper seemed to be in mm instead of cm, despite the factor RCD2H was described as being in cm^3. How exactly RCD2H was calculated from RCD and H was not clarified in the paper (units of measurement used).